dashboard/gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-jose==3.3.0
cryptography==42.0.5
pydantic==2.6.4
pydantic-settings==2.2.1
